pyproject.toml
[project]
name = "mpv2-panel-de-personas"
version = "3.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web Framework
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "werkzeug>=3.0.0",
    "gunicorn>=21.0.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
    "openpyxl>=3.1.0",
    
    # Azure Storage
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
extraction = [
    # Solo para extracción de datos (no necesario en producción)
    "scp>=0.14.0",
    "paramiko>=3.0.0",
    "pyodbc>=5.0.0",
    "sqlparse>=0.5.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
exclude = [
    "datos*",
    "frontend*",
    "tests*",
    "deploy*",
    "*.tests*",
]

[project.scripts]
panel-app = "main:main"

[tool.setuptools.package-data]
backend = ["*.xlsx", "data/*"]

